+++ extensions/cornerstone/src/services/ViewportService/CornerstoneViewportService.ts
@@ -300,7 +300,7 @@ class CornerstoneViewportService extends PubSubService implements IViewportServi
 
     return {
       viewportType: viewportInfo.getViewportType(),
-      viewReference: csViewport.getViewReference(),
+      viewReference: csViewport instanceof VolumeViewport3D ? null : csViewport.getViewReference(),
       position: csViewport.getViewPresentation({ pan: true, zoom: true }),
     };
   }
@@ -366,6 +366,14 @@ class CornerstoneViewportService extends PubSubService implements IViewportServi
     presentations?: Presentations
   ): void {
     const renderingEngine = this.getRenderingEngine();
+
+    // if not valid viewportData then return early
+    if (viewportData.viewportType === csEnums.ViewportType.STACK) {
+      // check if imageIds is valid
+      if (!viewportData.data[0].imageIds?.length) {
+        return;
+      }
+    }
 
     // This is the old viewportInfo, which may have old options but we might be
     // using its viewport (same viewportId as the new viewportInfo)