+++ extensions/cornerstone/src/utils/measurementServiceMappings/Probe.ts
@@ -23,7 +23,7 @@ const Probe = {
     const { metadata, data, annotationUID } = annotation;
 
     if (!metadata || !data) {
-      console.warn('Length tool: Missing metadata or data');
+      console.warn('Probe tool: Missing metadata or data');
       return null;
     }