+++ platform/core/src/classes/MetadataProvider.ts
@@ -8,31 +8,13 @@ import toNumber from '../utils/toNumber';
 import combineFrameInstance from '../utils/combineFrameInstance';
 
 class MetadataProvider {
-  constructor() {
-    // Define the main "metadataLookup" private property as an immutable property.
-    Object.defineProperty(this, 'studies', {
-      configurable: false,
-      enumerable: false,
-      writable: false,
-      value: new Map(),
-    });
-    Object.defineProperty(this, 'imageURIToUIDs', {
-      configurable: false,
-      enumerable: false,
-      writable: false,
-      value: new Map(),
-    });
-    // Can be used to store custom metadata for a specific type.
-    // For instance, the scaling metadata for PET can be stored here
-    // as type "scalingModule"
-    //
-    Object.defineProperty(this, 'customMetadata', {
-      configurable: false,
-      enumerable: false,
-      writable: false,
-      value: new Map(),
-    });
-  }
+  private readonly studies: Map<string, any> = new Map();
+  private readonly imageURIToUIDs: Map<string, any> = new Map();
+  private readonly imageUIDsByImageId: Map<string, any> = new Map();
+  // Can be used to store custom metadata for a specific type.
+  // For instance, the scaling metadata for PET can be stored here
+  // as type "scalingModule"
+  private readonly customMetadata: Map<string, any> = new Map();
 
   addImageIdToUIDs(imageId, uids) {
     // This method is a fallback for when you don't have WADO-URI or WADO-RS.
@@ -40,6 +22,7 @@ class MetadataProvider {
     // An example would be dicom hosted at some random site.
     const imageURI = imageIdToURI(imageId);
     this.imageURIToUIDs.set(imageURI, uids);
+    this.imageUIDsByImageId.set(imageId, uids);
   }
 
   addCustomMetadata(imageId, type, metadata) {
@@ -482,6 +465,12 @@ class MetadataProvider {
     //   SOPInstanceUID,
     // })
     // somewhere else
+
+    const cachedUIDs = this.imageUIDsByImageId.get(imageId);
+    if (cachedUIDs) {
+      return cachedUIDs;
+    }
+
     if (imageId.startsWith('wadors:')) {
       const strippedImageId = imageId.split('/studies/')[1];
       const splitImageId = strippedImageId.split('/');